requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-appconf==1.0.5
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-import-export==2.7.1
django-js-asset==1.2.2
django-mptt==0.13.4
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
pilkit==2.0
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
tablib==3.1.0
typing_extensions==4.1.1
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
